Create Day_5_Project_BMI_Calculator_v2.py
This is a v2 of the first BMI calculator. The main difference between this version compared to the first is that, this actually explains the BMI number. Also this uses if statements. I plan to go back to this project and add more and more to it just to see how far I've come along. Maybe for the next version I will have it in feet/pounds and add a conversion to metric.